P12-analytics/assetarc-analytics/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
python-dotenv==1.0.1
Jinja2==3.1.4
WeasyPrint==62.3
